[*] Added IdHelper and CustomData [*] Separated WeaponOverrideManager to its own class

+++ JellyLib/WeaponUtils/ProjectileOverride.cs
@@ -5,8 +5,8 @@
 /// </summary>
 public struct ProjectileOverride
 {
-    public int? damage;
-    public int? balanceDamage;
-    public int? explosionDamage;
-    public int? explosionBalanceDamage;
+    public float? damage;
+    public float? balanceDamage;
+    public float? explosionDamage;
+    public float? explosionBalanceDamage;
 }
